SEMANA07/dia4/pos_backend/requirements.txt
django==3.2
djangorestframework
mysqlclient
django-cors-headers
cloudinary